+++ .env.bsc
@@ -1,0 +1,1 @@
+ETHERSCAN_KEY=H8HVRCI8G15N51QPYKAM77T9XGPPW1SJKX
