+++ .env.eth
@@ -1,0 +1,1 @@
+ETHERSCAN_KEY=EH7MW2W9R7MXBW9E9V8YX6XGAQREMDHRTS
