+++ scripts/config_utils.ts
@@ -1,0 +1,32 @@
+export interface IConfig {
+  autoVerifyContract: boolean;
+
+  // mostly used for testing purpose
+  tokens: Array<{
+    symbol: string;
+    address: string;
+  }>;
+
+  // wrapped native token (wEth/wBnb ..)
+  wNative: string;
+
+  // Uniswap or clones
+  uniswap?: {
+    routers: string[];
+  };
+
+  // Compound or clones
+  compound?: {
+    compTroller: string;
+    cNative: string;
+    cTokens: string[];
+  };
+
+  supportedWallets: string[];
+}
+
+export const commonPlatformWallets = [
+  '0xa1738F8DD7c42cd4175CcdCa79Af89b3EC7b68E9', // android
+  '0x5250b8202AEBca35328E2c217C687E894d70Cd31', // ios
+  '0x168E4c3AC8d89B00958B6bE6400B066f0347DDc9', // web
+];
