+++ tsconfig.json
@@ -10,5 +10,5 @@
     "types": ["@nomiclabs/hardhat-ethers", "@nomiclabs/hardhat-waffle"]
   },
   "include": ["./test"],
-  "files": ["./hardhat.config.ts"]
+  "files": ["./hardhat.config.ts", "./scripts/IConfig.ts"]
 }
